Fix overflow errors in bitshift operations

+++ faucon-emu/src/memory/mod.rs
@@ -71,7 +71,7 @@ impl Memory {
     /// Reads a word from a given offset in a physical page of the code
     /// space.
     pub fn read_code_page(&self, page: u8, offset: u8) -> u32 {
-        self.read_code_addr(((page << 8) | offset) as u16)
+        self.read_code_addr(((page as u16) << 8) | offset as u16)
     }
 
     /// Reads a word from a given physical address in code space.
@@ -115,7 +115,7 @@ impl Memory {
     /// Writes a word to a given offset in a physical page of the code
     /// space.
     pub fn write_code_page(&mut self, page: u8, offset: u8, value: u32) {
-        self.write_code_addr(((page << 8) | offset) as u16, value);
+        self.write_code_addr(((page as u16) << 8) | offset as u16, value);
     }
 
     /// Writes a word to a given physical address in code space.
